Session4 code Rebasing
