+++ Session4/Group/Personnel.java
@@ -1,5 +1,3 @@
-package lk.ac.pdn;
-
 import java.security.SecureRandom;
 import java.util.Arrays;
 import java.util.Scanner;
